+++ spark-bigquery-dsv2/spark-3.1-bigquery/src/test/java/com/google/cloud/spark/bigquery/integration/Spark31ReadFromQueryIntegrationTest.java
@@ -16,7 +16,9 @@
 package com.google.cloud.spark.bigquery.integration;
 
 public class Spark31ReadFromQueryIntegrationTest extends ReadFromQueryIntegrationTestBase {
-
+  public Spark31ReadFromQueryIntegrationTest() {
+    super(true);
+  }
   // tests are from the super-class
 
 }