+++ spark-bigquery-dsv2/spark-3.2-bigquery/src/test/java/com/google/cloud/spark/bigquery/integration/Spark32ReadFromQueryIntegrationTest.java
@@ -17,6 +17,10 @@ package com.google.cloud.spark.bigquery.integration;
 
 public class Spark32ReadFromQueryIntegrationTest extends ReadFromQueryIntegrationTestBase {
 
+  public Spark32ReadFromQueryIntegrationTest() {
+    super(true);
+  }
+
   // tests are from the super-class
 
 }