+++ spark-bigquery-dsv2/spark-3.3-bigquery-lib/src/main/java/com/google/cloud/spark/bigquery/v2/Spark33BigQueryTable.java
@@ -15,17 +15,17 @@
  */
 package com.google.cloud.spark.bigquery.v2;
 
-import com.google.cloud.bigquery.TableId;
 import com.google.cloud.spark.bigquery.v2.context.BigQueryDataSourceReaderContext;
 import com.google.inject.Injector;
+import java.util.function.Supplier;
 import org.apache.spark.sql.connector.read.ScanBuilder;
 import org.apache.spark.sql.types.StructType;
 import org.apache.spark.sql.util.CaseInsensitiveStringMap;
 
 public class Spark33BigQueryTable extends Spark32BigQueryTable {
 
-  protected Spark33BigQueryTable(Injector injector, TableId tableId, StructType schema) {
-    super(injector, tableId, schema);
+  protected Spark33BigQueryTable(Injector injector, Supplier<StructType> schemaSupplier) {
+    super(injector, schemaSupplier);
   }
 
   @Override