+++ spark-bigquery-dsv2/spark-3.4-bigquery/src/test/java/com/google/cloud/spark/bigquery/integration/Spark34ReadFromQueryIntegrationTest.java
@@ -17,6 +17,10 @@ package com.google.cloud.spark.bigquery.integration;
 
 public class Spark34ReadFromQueryIntegrationTest extends ReadFromQueryIntegrationTestBase {
 
+  public Spark34ReadFromQueryIntegrationTest() {
+    super(true);
+  }
+
   // tests are from the super-class
 
 }