+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## Next
 
+* PR #1155: allow lazy materialization of query on load
+
 ## 0.35.1 - 2023-12-28
 
 * PR #1153: allow writing spark string to BQ datetime